fix: resolve compiler warnings and build errors
- Add __attribute__((unused)) to suppress unused parameter warnings
- Fix terminate function signature to match HDF5 VOL API expectation
- Update header declarations to match const function implementations
- Fix printf format specifier for hsize_t using cast to unsigned long
- Ensure consistent function signatures between header and source

+++ test/test_geotiff_read.c
@@ -71,7 +71,7 @@ int main(int argc, char **argv)
                 H5Sget_simple_extent_dims(space_id, dims, NULL);
                 printf("Image dimensions: ");
                 for (int i = 0; i < ndims; i++) {
-                    printf("%llu%s", dims[i], (i < ndims - 1) ? " x " : "");
+                    printf("%lu%s", (unsigned long) dims[i], (i < ndims - 1) ? " x " : "");
                 }
                 printf("\n");
             }
